+++ views/show.ejs
@@ -43,9 +43,6 @@
 						      <textarea class="form-control" id="exampleFormControlTextarea1" rows="3" name="data[text]"></textarea>
 						    </div>
 						    <div class="col-lg-2">
-						      <input type="text" class="form-control" placeholder="user" name="data[author]" value="<%= currentUser.username%>">
-						    </div>
-						    <div class="col-lg-2">
 						    	<button type="Submit" class="btn btn-primary">Submit</button>
 						    </div>
 						  </div>
